move aliasToSrc to module
